+++ pinot-plugins/pinot-batch-ingestion/pinot-batch-ingestion-standalone/src/main/java/org/apache/pinot/plugin/batch/ingestion/standalone/SegmentUriPushJobRunner.java
@@ -16,12 +16,10 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.pinot.plugin.ingestion.standalone;
+package org.apache.pinot.plugin.batch.ingestion.standalone;
 
-import com.google.common.base.Preconditions;
 import java.io.File;
 import java.io.IOException;
-import java.io.InputStream;
 import java.net.URI;
 import java.net.URISyntaxException;
 import java.util.ArrayList;
@@ -34,11 +32,11 @@ import org.apache.pinot.common.utils.FileUploadDownloadClient;
 import org.apache.pinot.common.utils.SimpleHttpResponse;
 import org.apache.pinot.spi.filesystem.PinotFS;
 import org.apache.pinot.spi.filesystem.PinotFSFactory;
-import org.apache.pinot.spi.ingestion.runner.IngestionJobRunner;
-import org.apache.pinot.spi.ingestion.spec.Constants;
-import org.apache.pinot.spi.ingestion.spec.PinotClusterSpec;
-import org.apache.pinot.spi.ingestion.spec.PinotFSSpec;
-import org.apache.pinot.spi.ingestion.spec.SegmentGenerationJobSpec;
+import org.apache.pinot.spi.batch.ingestion.runner.IngestionJobRunner;
+import org.apache.pinot.spi.batch.ingestion.spec.Constants;
+import org.apache.pinot.spi.batch.ingestion.spec.PinotClusterSpec;
+import org.apache.pinot.spi.batch.ingestion.spec.PinotFSSpec;
+import org.apache.pinot.spi.batch.ingestion.spec.SegmentGenerationJobSpec;
 import org.apache.pinot.spi.utils.retry.AttemptsExceededException;
 import org.apache.pinot.spi.utils.retry.RetriableOperationException;
 import org.apache.pinot.spi.utils.retry.RetryPolicies;
@@ -46,22 +44,25 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 
-public class SegmentTarPushJobRunner implements IngestionJobRunner {
+public class SegmentUriPushJobRunner implements IngestionJobRunner {
 
-  private static final Logger LOGGER = LoggerFactory.getLogger(SegmentTarPushJobRunner.class);
+  private static final Logger LOGGER = LoggerFactory.getLogger(SegmentUriPushJobRunner.class);
 
   private SegmentGenerationJobSpec _spec;
 
-  public SegmentTarPushJobRunner() {
+  public SegmentUriPushJobRunner() {
   }
 
-  public SegmentTarPushJobRunner(SegmentGenerationJobSpec spec) {
+  public SegmentUriPushJobRunner(SegmentGenerationJobSpec spec) {
     init(spec);
   }
 
   @Override
   public void init(SegmentGenerationJobSpec spec) {
     _spec = spec;
+    if (_spec.getPushJobSpec() == null) {
+      throw new RuntimeException("Missing PushJobSpec");
+    }
   }
 
   @Override
@@ -73,7 +74,7 @@ public class SegmentTarPushJobRunner implements IngestionJobRunner {
       PinotFSFactory.register(pinotFSSpec.getScheme(), pinotFSSpec.getClassName(), config);
     }
 
-    //Get outputFS for writing output pinot segments
+    //Get outputFS for writing output Pinot segments
     URI outputDirURI;
     try {
       outputDirURI = new URI(_spec.getOutputDirURI());
@@ -84,6 +85,7 @@ public class SegmentTarPushJobRunner implements IngestionJobRunner {
       throw new RuntimeException("outputDirURI is not valid - '" + _spec.getOutputDirURI() + "'");
     }
     PinotFS outputDirFS = PinotFSFactory.create(outputDirURI.getScheme());
+
     //Get list of files to process
     String[] files;
     try {
@@ -91,33 +93,29 @@ public class SegmentTarPushJobRunner implements IngestionJobRunner {
     } catch (IOException e) {
       throw new RuntimeException("Unable to list all files under outputDirURI - '" + outputDirURI + "'");
     }
-
-    List<String> segmentsToPush = new ArrayList<>();
+    List<String> segmentUris = new ArrayList<>();
     for (String file : files) {
-      if (file.endsWith(Constants.TAR_GZ_FILE_EXT)) {
-        segmentsToPush.add(file);
+      URI uri = URI.create(file);
+      if (uri.getPath().endsWith(Constants.TAR_GZ_FILE_EXT)) {
+        segmentUris.add(_spec.getPushJobSpec().getSegmentUriPrefix() + uri.getRawPath() + _spec.getPushJobSpec()
+            .getSegmentUriSuffix());
       }
     }
     try {
-      pushSegments(outputDirFS, segmentsToPush);
+      sendSegmentUris(segmentUris);
     } catch (RetriableOperationException | AttemptsExceededException e) {
       throw new RuntimeException(e);
     }
   }
 
-  public void pushSegments(PinotFS fileSystem, List<String> tarFilePaths)
+  public void sendSegmentUris(List<String> segmentUris)
       throws RetriableOperationException, AttemptsExceededException {
     String tableName = _spec.getTableSpec().getTableName();
-    LOGGER.info("Start pushing segments: {}... to locations: {} for table {}",
-        Arrays.toString(tarFilePaths.subList(0, Math.min(5, tarFilePaths.size())).toArray()),
-        Arrays.toString(_spec.getPinotClusterSpecs()), tableName);
+    LOGGER.info("Start sending table {} segment URIs: {} to locations: {}", tableName,
+        Arrays.toString(segmentUris.subList(0, Math.min(5, segmentUris.size())).toArray()),
+        Arrays.toString(_spec.getPinotClusterSpecs()));
     FileUploadDownloadClient fileUploadDownloadClient = new FileUploadDownloadClient();
-    for (String tarFilePath : tarFilePaths) {
-      URI tarFileURI = URI.create(tarFilePath);
-      File tarFile = new File(tarFilePath);
-      String fileName = tarFile.getName();
-      Preconditions.checkArgument(fileName.endsWith(Constants.TAR_GZ_FILE_EXT));
-      String segmentName = fileName.substring(0, fileName.length() - Constants.TAR_GZ_FILE_EXT.length());
+    for (String segmentUri : segmentUris) {
       for (PinotClusterSpec pinotClusterSpec : _spec.getPinotClusterSpecs()) {
         URI controllerURI;
         try {
@@ -125,7 +123,7 @@ public class SegmentTarPushJobRunner implements IngestionJobRunner {
         } catch (URISyntaxException e) {
           throw new RuntimeException("Got invalid controller uri - '" + pinotClusterSpec.getControllerURI() + "'");
         }
-        LOGGER.info("Pushing segment: {} to location: {} for table {}", segmentName, controllerURI, tableName);
+        LOGGER.info("Sending table {} segment URI: {} to location: {} for ", tableName, segmentUri, controllerURI);
         int attempts = 1;
         if (_spec.getPushJobSpec() != null && _spec.getPushJobSpec().getPushAttempts() > 0) {
           _spec.getPushJobSpec().getPushAttempts();
@@ -135,25 +133,23 @@ public class SegmentTarPushJobRunner implements IngestionJobRunner {
           retryWaitMs = _spec.getPushJobSpec().getPushRetryIntervalMillis();
         }
         RetryPolicies.exponentialBackoffRetryPolicy(attempts, retryWaitMs, 5).attempt(() -> {
-          try (InputStream inputStream = fileSystem.open(tarFileURI)) {
+          try {
             SimpleHttpResponse response = fileUploadDownloadClient
-                .uploadSegment(FileUploadDownloadClient.getUploadSegmentURI(controllerURI), segmentName, inputStream,
-                    tableName);
-            LOGGER.info("Response for pushing table {} segment {} to location {} - {}: {}", tableName, segmentName,
+                .sendSegmentUri(FileUploadDownloadClient.getUploadSegmentURI(controllerURI), segmentUri, tableName);
+            LOGGER.info("Response for pushing table {} segment uri {} to location {} - {}: {}", tableName, segmentUri,
                 controllerURI, response.getStatusCode(), response.getResponse());
             return true;
           } catch (HttpErrorStatusException e) {
             int statusCode = e.getStatusCode();
             if (statusCode >= 500) {
               // Temporary exception
-              LOGGER.warn("Caught temporary exception while pushing table: {} segment: {} to {}, will retry", tableName,
-                  segmentName, controllerURI, e);
+              LOGGER.warn("Caught temporary exception while pushing table: {} segment uri: {} to {}, will retry",
+                  tableName, segmentUri, controllerURI, e);
               return false;
             } else {
               // Permanent exception
-              LOGGER
-                  .error("Caught permanent exception while pushing table: {} segment: {} to {}, won't retry", tableName,
-                      segmentName, controllerURI, e);
+              LOGGER.error("Caught permanent exception while pushing table: {} segment uri: {} to {}, won't retry",
+                  tableName, segmentUri, controllerURI, e);
               throw e;
             }
           }
