+++ pinot-spi/src/main/java/org/apache/pinot/spi/batch/ingestion/spec/SegmentGenerationTaskSpec.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.pinot.spi.ingestion.spec;
+package org.apache.pinot.spi.batch.ingestion.spec;
 
 import com.fasterxml.jackson.databind.JsonNode;
 import org.apache.pinot.spi.data.Schema;
