+++ pinot-tools/pom.xml
@@ -288,7 +288,7 @@
               </jvmSettings>
             </program>
             <program>
-              <mainClass>org.apache.pinot.spi.ingestion.IngestionJobLauncher</mainClass>
+              <mainClass>org.apache.pinot.spi.batch.ingestion.IngestionJobLauncher</mainClass>
               <name>pinot-ingestion-job</name>
               <jvmSettings>
                 <initialMemorySize>1G</initialMemorySize>
